meta: move thefourtheye to TSC Emeritus
thefourtheye has a considerable history of contributions to Node.js.
They have not been active much of late, and the TSC Charter has a
section about activity indicating that moving to Emeritus at this time
is the thing to do. Thanks for all you've done to make Node.js
fantastic, thefourtheye, and hope to see you around again soon!

PR-URL: https://github.com/nodejs/node/pull/32059
Reviewed-By: Anna Henningsen <anna@addaleax.net>
Reviewed-By: Matteo Collina <matteo.collina@gmail.com>
Reviewed-By: Colin Ihrig <cjihrig@gmail.com>
Reviewed-By: Anatoli Papirovski <apapirovski@mac.com>
Reviewed-By: Joyee Cheung <joyeec9h3@gmail.com>
Reviewed-By: Luigi Pinca <luigipinca@gmail.com>
Reviewed-By: Michael Dawson <michael_dawson@ca.ibm.com>
Reviewed-By: Tobias Nießen <tniessen@tnie.de>
Reviewed-By: Сковорода Никита Андреевич <chalkerx@gmail.com>